Beta_003 version 0.8.59 in media res
about 0.8.56 is stable but i didnt commit. so the script versions in the ms access (good one) file are the most stable versions. here i'm in the middle of changing insertmrllineitems and executeprotectedfuntion to accommodate larger batches, but i should've committed but didn't. there are alpha versions of those in vbcode which may not work.